// apps/pwa/src/types/domain.ts
export type DashboardCards = {
  todaysSales: number;
  todaysProfit: number;
  todaysExpenses: number;
  todaysOrders: number;
  totalInvestment: number;
  totalRevenue: number;
  totalProfit: number;
  netProfit: number;
  inventoryValue: number;
  remainingStock: number;
  lowStockCount: number;
  outOfStockCount: number;
};